requirements.txt
Flask==2.3.3
paho-mqtt==1.6.1
pymongo==4.5.0
